+++ change-log.md
@@ -5,6 +5,7 @@
 - Fixed detect when compound statement closer is not needed (#50)
 - Fixed avoid de-indenting compound statement closers beyond opener level (#48)
 - Fixed single-line `asm` statements no longer treated as multi-line block openers (#46)
+- Fixed auto-indent and code formatter no longer add `End Asm` closer / indent body for single-line `asm` (#46)
 
 # Changes since 0.5.0-beta.2